libOpenCL: Require ocl-icd or OpenCL-ICD-Loader (RHBZ#2332429)
Since introducing OpenCL-ICD-Loader as a new package, an alternative to ocl-icd,
Fedora composes under some circumstances do pick the OpenCL-ICD-Loader instead of
ocl-icd due to something requiring the .so file, and not the ocl-icd specifically.

This change allows the mesa-libOpenCL to work with both ocl icd implementations,
and workarounds FTI issue that occurs if OpenCL-ICD-Loader is present, and not ocl-icd.