Fix failure for Asahi disabled in RHEL builds
The Apple DRI driver is part of the KMSRO framework and is always
built when any KMSRO driver is enabled. This is expressed as a
symlink file named "apple_dri.so" pointing to the main driver file.

When we aren't shipping the asahi driver, we do not need this file,
so delete it in that case.

Fixes: 7911753afd8fd96866fd5a933b7720084f31d118 ("Disable asahi in RHEL builds")